+++ src/com/tombloom/game/GameState.java
@@ -1,7 +1,7 @@
 package com.tombloom.game;
 
 public class GameState {
-    public static int health;
+    public static float health;
     public static int score = 0;
     public static int level = 1;
 
@@ -10,7 +10,7 @@ public class GameState {
     }
 
     public void tick(){
-        setHealth(Game.clamp(health, 0, 100));
+        setHealth(Game.clamp(health, 0f, 100f));
         score++;
     }
 
@@ -18,7 +18,7 @@ public class GameState {
         health -= amount;
     }
 
-    public static void setHealth(int value){
+    public static void setHealth(float value){
         health = value;
     }
 }
